refactor: move handlers to new package
